\name{hmyriB36}
\alias{hmyriB36}
\title{hmyriB36 -- representation of Yoruba hapmap data, build 36 (r23a)  }
\description{hmyriB36 -- representation of Yoruba hapmap data, build 36 (r23a)  }
\usage{
data(hmyriB36)
}
\details{
}
\value{
instance of GGBase smlSet class
}
\references{\url{www.hapmap.org}}
\author{Vince Carey <stvjc@channing.harvard.edu>}
%\note{  }
%
%
%\seealso{  }
%
\examples{
library(GGtools)
getSS("hmyriB36", "20")
}
\keyword{ models }
